src/data/markets.py
"""
ORACLE-26 — Prediction Markets Data
=====================================
Fetches real-money betting odds and implied probabilities.
Sources: 
  1. Polymarket (Public API)
  2. Kalshi (Authenticated API)
"""

import os
import requests
import pandas as pd
from pathlib import Path
from dotenv import load_dotenv

import sys
sys.path.insert(0, str(Path(__file__).resolve().parents[2]))

from src.constants import ALL_TEAMS

load_dotenv()

RAW_DIR = Path("data/raw")
RAW_DIR.mkdir(parents=True, exist_ok=True)

class MarketDataCollector:
    """
    Pulls prediction market data to derive implied probabilities.
    """

    def __init__(self):
        self.kalshi_email = os.getenv("KALSHI_EMAIL")
        self.kalshi_pass  = os.getenv("KALSHI_PASSWORD")

    def fetch_polymarket(self):
        """Fetch World Cup winner odds from Polymarket."""
        print("  [markets] Fetching Polymarket odds...")
        # Note: In a real scenario, we'd search for the specific WC2026 market ID.
        # Here we'll simulate the response based on their API structure.
        
        market_data = []
        for team in ALL_TEAMS:
            # Simulated implied probability (to be replaced with live API call)
            # In Phase 3, we'll implement the actual GraphQL/REST search logic.
            implied_prob = 0.02 # Placeholder
            market_data.append({"team": team, "source": "polymarket", "prob": implied_prob})
            
        df = pd.DataFrame(market_data)
        return df

    def fetch_kalshi(self):
        """Fetch odds from Kalshi."""
        print("  [markets] Fetching Kalshi odds (Auth required)...")
        # Kalshi API implementation here
        return pd.DataFrame()

    def run(self):
        print("\n" + "="*50)
        print("📈 ORACLE-26 | Phase 2.1 — Market Signals")
        print("="*50)
        poly_df = self.fetch_polymarket()
        kalshi_df = self.fetch_kalshi()
        
        combined = pd.concat([poly_df, kalshi_df])
        out_path = RAW_DIR / "market_signals.csv"
        combined.to_csv(out_path, index=False)
        print(f"  [markets] Saved market signals to {out_path}")

if __name__ == "__main__":
    collector = MarketDataCollector()
    collector.run()
