requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alpha_vantage==3.0.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
behave==1.2.6
blinker==1.9.0
branca==0.8.2
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cssselect==1.4.0
cycler==0.12.1
exceptiongroup==1.3.1
execnet==2.1.2
fastapi==0.136.1
fasteners==0.20
filelock==3.29.0
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.62.1
fredapi==0.5.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
html5lib==1.1
idna==3.13
iniconfig==2.3.0
jh2==5.0.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llvmlite==0.47.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.7.1
mycdp==1.3.7
narwhals==2.20.0
niquests==3.18.6
numba==0.65.0
numpy==2.4.4
nvidia-nccl-cu12==2.30.4
openmeteo_requests==1.7.5
openmeteo_sdk==1.26.0
orjson==3.11.8
outcome==1.3.0.post0
packaging==26.1
pandas==3.0.2
parameterized==0.9.0
parse==1.21.1
parse_type==0.6.6
pdbp==1.8.2
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==24.0.0
PyAutoGUI==0.9.54
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pynose==1.5.5
pyotp==2.9.0
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==9.0.3
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-xlib==0.33
python3-xlib==0.15
pytrends==4.9.2
pytweening==1.2.0
PyYAML==6.0.3
qh3==1.7.3
referencing==0.37.0
requests==2.33.1
retry-requests==2.0.0
rich==15.0.0
rpds-py==0.30.0
sbvirtualdisplay==1.4.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
selenium==4.43.0
seleniumbase==4.48.2
setuptools==82.0.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
smmap==5.0.3
sniffio==1.3.1
soccerdata==1.9.0
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==1.0.0
streamlit==1.56.0
tabcompleter==1.4.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
update-checker==0.18.0
urllib3==2.6.3
urllib3-future==2.19.910
uvicorn==0.46.0
wassima==2.0.6
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.47.0
wrapper-tls-requests==1.2.5
wsproto==1.3.2
xgboost==3.2.0
xyzservices==2026.3.0
yarl==1.23.0
